chore: move src/server/fetcher.ts into src/server/fetcher/ folder
