+++ src/server/auth-client.ts
@@ -192,7 +192,7 @@ import {
   FetcherConfig,
   FetcherHooks,
   FetcherMinimalConfig
-} from "./fetcher.js";
+} from "./fetcher/fetcher.js";
 import {
   buildForwardedRequestHeaders,
   buildForwardedResponseHeaders,
