+++ src/server/fetcher/fetcher.msw.test.ts
@@ -15,7 +15,7 @@ import {
 import { createNextHeadersMock } from "../../test-fixtures/mocks.js";
 import { SessionData } from "../../types/index.js";
 import { Auth0Client } from "../client.js";
-import { Fetcher } from "../fetcher.js";
+import { Fetcher } from "./fetcher.js";
 
 vi.mock("next/headers.js", () => createNextHeadersMock());
